# testFuncional/swagger-persona.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"#persona","description":"Persona Controller"}],"paths":{"/persona/actualizar":{"put":{"tags":["#persona"],"summary":"Actualizar Persona","operationId":"actualizarPersonaUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"persona","description":"persona","required":true,"schema":{"$ref":"#/definitions/Persona"}}],"responses":{"200":{"description":"Persona actualizada."},"201":{"description":"Created"},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/persona/consultar":{"get":{"tags":["#persona"],"summary":"Consultar Personas","operationId":"consultarPersonasUsingGET","produces":["*/*"],"responses":{"200":{"description":"Personas consultadas.","schema":{"type":"array","items":{"$ref":"#/definitions/Persona"}}},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/persona/consultar/{id}":{"get":{"tags":["#persona"],"summary":"Consultar Persona por ID","operationId":"consultarPersonaPorIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Persona consultada.","schema":{"$ref":"#/definitions/Persona"}},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/persona/eliminar/{id}":{"delete":{"tags":["#persona"],"summary":"Eliminar Persona","operationId":"eliminarPersonaUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"Persona eliminada."},"204":{"description":"No Content"},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"500":{"description":"Error inesperado."}},"deprecated":false}},"/persona/guardar":{"post":{"tags":["#persona"],"summary":"Crear Persona","operationId":"guardarPersonaUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"persona","description":"persona","required":true,"schema":{"$ref":"#/definitions/Persona"}}],"responses":{"200":{"description":"Persona creada.","schema":{"$ref":"#/definitions/Persona"}},"201":{"description":"Created"},"400":{"description":"Datos de entrada incorrectos."},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"500":{"description":"Error inesperado."}},"deprecated":false}}},"definitions":{"Persona":{"type":"object","required":["apellidos","correo","nombres","numDocumento","tipoDocumento"],"properties":{"apellidos":{"type":"string","description":"Apellidos de la persona"},"correo":{"type":"string","description":"Correo de la persona"},"id":{"type":"integer","format":"int64"},"nombres":{"type":"string","description":"Nombres de la persona"},"numDocumento":{"type":"integer","format":"int64","description":"Numero de documento de la persona"},"tipoDocumento":{"type":"string","description":"Tipo de documento de la persona"}},"title":"Persona"}}}